# Avoid some allocations in recvMsg