@dsymonds 2015-10-05
Nice gain for relatively little code impact.

@iamqizhao: the go/ links on the types above this function really need public versions!