@iamqizhao 2015-10-05
On Mon, Oct 5, 2015 at 2:24 PM, David Symonds <notifications@github.com>
wrote:

> Nice gain for relatively little code impact.
>
> @iamqizhao <https://github.com/iamqizhao>: the go/ links on the types
> above this function really need public versions!
>
sure, will take care of it shortly.

> —
> Reply to this email directly or view it on GitHub
> <https://github.com/grpc/grpc-go/pull/385#issuecomment-145672647>.
>