@iamqizhao 2015-10-06
I saw one outlier
BenchmarkClientUnaryc1-8 223208 257968 +15.57%

can you rerun the benchmark a couple of times (or longer time) to verify?