@nuss-justin 2015-10-06
Added constants and reran tests with `go test -bench . -benchmem -benchtime 3s`. Compared commits are 3e7b7e58f49 and 5165699f376. New results:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkClientStreamc1-8              121497        128525        +5.78%
BenchmarkClientStreamc8-8              20963         20454         -2.43%
BenchmarkClientStreamc64-8             12370         11980         -3.15%
BenchmarkClientStreamc512-8            13542         13152         -2.88%
BenchmarkClientUnaryc1-8               240343        237528        -1.17%
BenchmarkClientUnaryc8-8               46075         44415         -3.60%
BenchmarkClientUnaryc64-8              32634         33386         +2.30%
BenchmarkClientUnaryc512-8             33065         32732         -1.01%
BenchmarkClientStreamNoTracec1-8       115481        93490         -19.04%
BenchmarkClientStreamNoTracec8-8       20309         20031         -1.37%
BenchmarkClientStreamNoTracec64-8      11684         11413         -2.32%
BenchmarkClientStreamNoTracec512-8     12805         12218         -4.58%
BenchmarkClientUnaryNoTracec1-8        226422        221111        -2.35%
BenchmarkClientUnaryNoTracec8-8        41646         41079         -1.36%
BenchmarkClientUnaryNoTracec64-8       29446         28662         -2.66%
BenchmarkClientUnaryNoTracec512-8      29964         29044         -3.07%

benchmark                              old allocs     new allocs     delta
BenchmarkClientStreamc1-8              46             41             -10.87%
BenchmarkClientStreamc8-8              47             41             -12.77%
BenchmarkClientStreamc64-8             46             41             -10.87%
BenchmarkClientStreamc512-8            46             40             -13.04%
BenchmarkClientUnaryc1-8               100            91             -9.00%
BenchmarkClientUnaryc8-8               101            92             -8.91%
BenchmarkClientUnaryc64-8              100            91             -9.00%
BenchmarkClientUnaryc512-8             100            91             -9.00%
BenchmarkClientStreamNoTracec1-8       43             37             -13.95%
BenchmarkClientStreamNoTracec8-8       43             37             -13.95%
BenchmarkClientStreamNoTracec64-8      43             37             -13.95%
BenchmarkClientStreamNoTracec512-8     42             36             -14.29%
BenchmarkClientUnaryNoTracec1-8        87             78             -10.34%
BenchmarkClientUnaryNoTracec8-8        88             78             -11.36%
BenchmarkClientUnaryNoTracec64-8       87             78             -10.34%
BenchmarkClientUnaryNoTracec512-8      87             78             -10.34%

benchmark                              old bytes     new bytes     delta
BenchmarkClientStreamc1-8              2015          1920          -4.71%
BenchmarkClientStreamc8-8              2046          1950          -4.69%
BenchmarkClientStreamc64-8             2044          1948          -4.70%
BenchmarkClientStreamc512-8            2053          1957          -4.68%
BenchmarkClientUnaryc1-8               5983          5838          -2.42%
BenchmarkClientUnaryc8-8               5985          5841          -2.41%
BenchmarkClientUnaryc64-8              6010          5866          -2.40%
BenchmarkClientUnaryc512-8             5949          5809          -2.35%
BenchmarkClientStreamNoTracec1-8       1887          1791          -5.09%
BenchmarkClientStreamNoTracec8-8       1917          1821          -5.01%
BenchmarkClientStreamNoTracec64-8      1928          1832          -4.98%
BenchmarkClientStreamNoTracec512-8     1923          1828          -4.94%
BenchmarkClientUnaryNoTracec1-8        4303          4159          -3.35%
BenchmarkClientUnaryNoTracec8-8        4304          4160          -3.35%
BenchmarkClientUnaryNoTracec64-8       4330          4186          -3.33%
BenchmarkClientUnaryNoTracec512-8      4280          4139          -3.29%
```

Output of [benchstat](https://godoc.org/rsc.io/benchstat) (3 runs for each commit):

```
name                       old time/op    new time/op    delta
ClientStreamc1-8              122µs ± 1%     127µs ± 3%   ~     (p=0.200 n=3+3)
ClientStreamc8-8             21.0µs ± 0%    20.5µs ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc64-8            12.3µs ± 0%    12.0µs ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc512-8           13.5µs ± 0%    13.1µs ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc1-8               242µs ± 1%     237µs ± 1%   ~     (p=0.100 n=3+3)
ClientUnaryc8-8              45.2µs ± 2%    44.5µs ± 0%   ~     (p=0.400 n=3+3)
ClientUnaryc64-8             32.6µs ± 1%    32.6µs ± 2%   ~     (p=0.700 n=3+3)
ClientUnaryc512-8            32.9µs ± 0%    32.6µs ± 1%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec1-8       114µs ± 1%     105µs ±11%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec8-8      20.3µs ± 0%    19.8µs ± 1%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec64-8     11.6µs ± 1%    11.4µs ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec512-8    12.7µs ± 1%    12.2µs ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec1-8        228µs ± 1%     226µs ± 3%   ~     (p=0.700 n=3+3)
ClientUnaryNoTracec8-8       41.5µs ± 0%    41.1µs ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec64-8      29.4µs ± 0%    28.7µs ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec512-8     29.8µs ± 1%    29.0µs ± 1%   ~     (p=0.100 n=3+3)

name                       old alloc/op   new alloc/op   delta
ClientStreamc1-8             2.02kB ± 0%    1.92kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc8-8             2.05kB ± 0%    1.95kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc64-8            2.04kB ± 0%    1.95kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc512-8           2.05kB ± 0%    1.96kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc1-8              5.98kB ± 0%    5.84kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc8-8              5.99kB ± 0%    5.84kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc64-8             6.01kB ± 0%    5.87kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc512-8            5.95kB ± 0%    5.81kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec1-8      1.89kB ± 0%    1.79kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec8-8      1.92kB ± 0%    1.82kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec64-8     1.93kB ± 0%    1.83kB ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec512-8    1.92kB ± 0%    1.83kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec1-8       4.30kB ± 0%    4.16kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec8-8       4.30kB ± 0%    4.16kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec64-8      4.33kB ± 0%    4.19kB ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec512-8     4.28kB ± 0%    4.14kB ± 0%   ~     (p=0.100 n=3+3)

name                       old allocs/op  new allocs/op  delta
ClientStreamc1-8               46.7 ± 1%      41.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc8-8               47.0 ± 0%      41.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc64-8              46.3 ± 1%      41.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamc512-8             46.0 ± 0%      40.0 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc1-8                 100 ± 0%        91 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc8-8                 101 ± 0%        92 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc64-8                100 ± 0%        91 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryc512-8               100 ± 0%        91 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec1-8        43.0 ± 0%      37.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec8-8        43.0 ± 0%      37.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec64-8       42.7 ± 2%      37.0 ± 0%   ~     (p=0.100 n=3+3)
ClientStreamNoTracec512-8      42.0 ± 0%      36.0 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec1-8         87.0 ± 0%      78.0 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec8-8         88.0 ± 0%      78.7 ± 1%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec64-8        87.0 ± 0%      78.0 ± 0%   ~     (p=0.100 n=3+3)
ClientUnaryNoTracec512-8       87.0 ± 0%      78.0 ± 0%   ~     (p=0.100 n=3+3)
```